Move desktop server exposure into backend
- Relocate server exposure modules under backend
- Update imports and affected tests
